Unifies REST and SOAP clients into single interface
