+++ lib/zuora.rb
@@ -28,6 +28,7 @@ require_relative 'zuora/errors'
 require_relative 'zuora/utils/envelope'
 require_relative 'zuora/client'
 require_relative 'zuora/rest'
+require_relative 'zuora/soap'
 require_relative 'zuora/object'
 require_relative 'zuora/dispatcher'
 require_relative 'zuora/response'
