+++ lib/zuora/soap.rb
@@ -1,0 +1,6 @@
+module Zuora
+  module Soap
+  end
+end
+
+require_relative 'soap/client'
