+++ spec/cassettes/rest/auth_failure.yml
@@ -30,20 +30,17 @@ http_interactions:
       - Zuora App
       Content-Type:
       - application/json;charset=utf-8
-      X-N:
-      - S
+      Content-Length:
+      - '138'
       Expires:
-      - Wed, 10 Feb 2016 15:11:35 GMT
+      - Fri, 19 Feb 2016 15:53:45 GMT
       Cache-Control:
       - max-age=0, no-cache, no-store
       Pragma:
       - no-cache
       Date:
-      - Wed, 10 Feb 2016 15:11:35 GMT
-      Transfer-Encoding:
-      - chunked
+      - Fri, 19 Feb 2016 15:53:45 GMT
       Connection:
-      - Transfer-Encoding
       - close
     body:
       encoding: UTF-8
@@ -56,5 +53,5 @@ http_interactions:
           } ]
         }
     http_version: 
-  recorded_at: Wed, 10 Feb 2016 15:11:35 GMT
+  recorded_at: Fri, 19 Feb 2016 15:53:45 GMT
 recorded_with: VCR 3.0.1
