+++ spec/cassettes/soap/amend_add_product_success.yml
