+++ spec/cassettes/soap/amend_remove_product_success.yml
