+++ spec/cassettes/soap/amend_update_product_success.yml
