+++ spec/cassettes/soap/authentication_failure.yml
@@ -34,13 +34,13 @@ http_interactions:
       Content-Type:
       - text/xml;charset=UTF-8
       Expires:
-      - Thu, 18 Feb 2016 15:06:28 GMT
+      - Fri, 19 Feb 2016 15:53:40 GMT
       Cache-Control:
       - max-age=0, no-cache, no-store
       Pragma:
       - no-cache
       Date:
-      - Thu, 18 Feb 2016 15:06:28 GMT
+      - Fri, 19 Feb 2016 15:53:40 GMT
       Connection:
       - close
     body:
@@ -50,5 +50,5 @@ http_interactions:
         login. User name and password do not match.</faultstring><detail><fns:LoginFault><fns:FaultCode>INVALID_VALUE</fns:FaultCode><fns:FaultMessage>Invalid
         login. User name and password do not match.</fns:FaultMessage></fns:LoginFault></detail></soapenv:Fault></soapenv:Body></soapenv:Envelope>
     http_version: 
-  recorded_at: Thu, 18 Feb 2016 15:06:29 GMT
+  recorded_at: Fri, 19 Feb 2016 15:53:40 GMT
 recorded_with: VCR 3.0.1
