+++ spec/cassettes/soap/query_less_than_sign_success.yml
