+++ spec/cassettes/soap/query_simple_get_success.yml
